修正

+++ authoritymanage/src/main/java/com/smartindustry/authority/service/impl/DeptServiceImpl.java
@@ -77,7 +77,9 @@ public class DeptServiceImpl implements IDeptService {
         if (dids.contains(1L)){
             dids.remove(1L);
         }
-        deptMapper.deleteBatch(dids);
+        if(dids.size()!=0){
+            deptMapper.deleteBatch(dids);
+        }
         return ResultVO.ok();
     }
